src/requirements.txt
django==3.1.5
djangorestframework==3.12.2
psycopg2-binary>=2.8
django-cors-headers==3.7.0
djangorestframework-simplejwt
drf-nested-routers==0.93.*
coverage==5.5